Cart item rendering fixed, Cart empty feature fixed, checkout page reloading empty cart fixed
