+++ oms-client/src/Components/Header/Header.js
@@ -28,7 +28,6 @@ function Header() {
   const [showLogout, setShowLogout] = useState(false);
   const navigate = useNavigate();
   useEffect(() => {
-    
     if (localStorage.getItem("userDetails")) {
       setUserDetails(JSON.parse(localStorage.getItem("userDetails")));
     }
@@ -44,6 +43,36 @@ function Header() {
     // getCart();
     return;
   }, []);
+  const getCart = async () => {
+    const response = await fetch(
+      `${BASE_URL}/addtocart/getCartsByUserId/${
+        JSON.parse(localStorage.getItem("userDetails")).id
+      }`,
+      {
+        method: "GET",
+        headers: {
+          Authorization: `Bearer ${JSON.parse(localStorage.getItem("token"))}`,
+        },
+      }
+    );
+    const data = await response.json();
+    // console.log(data);
+    var temp = [];
+    if (data.length > 0) {
+      data.map((item) => {
+        let product = productHome.filter(
+          (medicine) => medicine.id === item.productId
+        )[0];
+        product = { ...product, quantity: item.quantity };
+        temp = [...temp, product];
+      });
+    }
+    setCart(temp);
+    // console.log(temp);
+  };
+  useEffect(() => {
+    getCart();
+  }, [productHome]);
 
   const handleLogout = () => {
     localStorage.removeItem("token");
