+++ oms-client/src/Pages/Checkout/Checkout.js
@@ -1,17 +1,19 @@
-import React from 'react'
-import CheckoutCard from '../../Components/CheckoutCards/checkoutcard/CheckoutCard';
-import CheckoutcartEmpty from '../../Components/CheckoutCards/checkoutcartEmpty/CheckoutcartEmpty'
+import React, { useContext, useEffect, useState } from "react";
+import CheckoutCard from "../../Components/CheckoutCards/checkoutcard/CheckoutCard";
+import CheckoutcartEmpty from "../../Components/CheckoutCards/checkoutcartEmpty/CheckoutcartEmpty";
+import { Cart } from "../../App";
 
 function Checkout() {
-  const isCartEmpty = false;
+  const [cart, setCart] = useContext(Cart);
+  // useEffect(() => {},[])
+  const [temp, setTemp] = useState(false);
+  useEffect(() => {
+    setTemp(true);
+  },[cart])
   
   return (
-    <div>
-      {
-          (isCartEmpty) ? <CheckoutcartEmpty /> : <CheckoutCard />
-      }
-    </div>
-  )
+    <div>{cart.length == 0 ? <CheckoutcartEmpty /> : <CheckoutCard />}</div>
+  );
 }
 
 export default Checkout;
